Delete generic telegram content type
It was initially fun to do, but implementing a bunch of other
necessary traits was more cumbersome than I expected.

Now instead, we divide the telegram into a base part and a data part.

This data part then consists of a mutually exclusive electric and gas
part.

Furthermore, the primitives, which consists of three values, now act the same way: after validation
only one of the string, float or date is guarantueed to exist.

Note: I may implement the primitives as enums instead, but time will tell.